pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "reproducible_sfc_not_working"
description = "Reproducible sfc not working"
version = "0.1.0"
readme = "README.md"
authors = ["blueshoe.de"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"

django = "^5.0"
django-hurricane = "^1.5.0"
django-split-settings = "^1.3.0"
django-csp = "^3.8"
django-http-referrer-policy = "^1.1"
django-permissions-policy = "^4.19"
django-extensions = "^3.2.3"
django-storages = "^1.14.2"

python-decouple = "^3.8"
bcrypt = "^4.1.2"
structlog = "^24.1"
django-compressor = "^4.4"
django-browser-reload = "^1.12.1"
django-components = "^0.64"


[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.2"
django-querycount = "^0.7"
django-coverage-plugin = "^2.0"

flake8-pytest-style = "^1.5"
flake8-django = "^1.1"
flake8-logging-format = "^0.6"

pytest = "^8.1.1"
pytest-django = "^4.8"
pytest-cov = "^2.12"
pytest-randomly = "^3.8"
pytest-deadfixtures = "^2.2"
pytest-testmon = "^2.1.1"
pytest-timeout = "^1.4"
django-test-migrations = "^1.1"
hypothesis = "^6.14"

mypy = "^1.9.0"

tomlkit = "^0.7"
doc8 = "^0.8"

yamllint = "^1.26"
safety = "^1.10"
polint = "^0.4"
black = "^24.0"

[tool.black]
line-length = 120
